Update error table to be persistent

+++ scripts/collector/sqlserver/sql/columnDatatypes.sql
@@ -57,8 +57,8 @@ CREATE TABLE #columnDatatypes(
    ,column_count nvarchar(255)
    );
 
-IF OBJECT_ID('tempdb..#dmaCollectorErrors') IS NULL 
-   CREATE TABLE #dmaCollectorErrors(
+IF OBJECT_ID('tempdb.dbo.dmaCollectorErrors') IS NULL 
+   CREATE TABLE tempdb.dbo.dmaCollectorErrors(
       database_name nvarchar(255) DEFAULT db_name()
       ,module_name nvarchar(255)
       ,error_number nvarchar(255)
@@ -194,7 +194,7 @@ BEGIN
       END;
    END TRY
    BEGIN CATCH
-      INSERT INTO #dmaCollectorErrors
+      INSERT INTO tempdb.dbo.dmaCollectorErrors
       SELECT
          db_name(),
          'columnDatatypes',
@@ -218,4 +218,5 @@ DEALLOCATE db_cursor
 
 SELECT @PKEY as PKEY, a.* from #columnDatatypes a;
 
-DROP TABLE #columnDatatypes;
+IF OBJECT_ID('tempdb..#columnDatatypes') IS NOT NULL  
+   DROP TABLE #columnDatatypes;
